refactor: portfolio to include async & extra functionality
- Better auth check for portfolio.
- Async for portfolio transactions.
- Refactored for conciseness.

+++ src/main/java/com/aston/stockapp/domain/portfolio/PortfolioStock.java
@@ -18,5 +18,5 @@ public class PortfolioStock {
 
     private String name;
 
-    private int currentPrice;
+    private double currentPrice;
 }
